+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.9
 
 * improve index templates, fixes [#242](https://github.com/cars10/elasticvue/issues/242)
+* improve shards table, fixes [#237](https://github.com/cars10/elasticvue/issues/237)
 * dependency updates
 
 ## 1.0.8